Moving all Bigtable row filters into their own module.
In the process also lowering the line limits to levels
before they were raised to accomodate the final changes
to the Bigtable Row implementations.
