+++ gcloud/bigquery/test_job.py
@@ -1,4 +1,3 @@
-# pylint: disable=C0302
 # Copyright 2015 Google Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
