+++ gcloud/bigtable/happybase/test_table.py
@@ -1113,7 +1113,7 @@ class Test__convert_to_time_range(unittest2.TestCase):
 
     def test_success(self):
         from gcloud._helpers import _datetime_from_microseconds
-        from gcloud.bigtable.row import TimestampRange
+        from gcloud.bigtable.row_filters import TimestampRange
 
         timestamp = 1441928298571
         ts_dt = _datetime_from_microseconds(1000 * timestamp)
@@ -1219,7 +1219,7 @@ class Test__filter_chain_helper(unittest2.TestCase):
             self._callFUT()
 
     def test_single_filter(self):
-        from gcloud.bigtable.row import CellsColumnLimitFilter
+        from gcloud.bigtable.row_filters import CellsColumnLimitFilter
 
         versions = 1337
         result = self._callFUT(versions=versions)
@@ -1229,7 +1229,7 @@ class Test__filter_chain_helper(unittest2.TestCase):
         self.assertEqual(result.num_cells, versions)
 
     def test_existing_filters(self):
-        from gcloud.bigtable.row import CellsColumnLimitFilter
+        from gcloud.bigtable.row_filters import CellsColumnLimitFilter
 
         filters = []
         versions = 1337
@@ -1244,9 +1244,9 @@ class Test__filter_chain_helper(unittest2.TestCase):
 
     def _column_helper(self, num_filters, versions=None, timestamp=None,
                        column=None, col_fam=None, qual=None):
-        from gcloud.bigtable.row import ColumnQualifierRegexFilter
-        from gcloud.bigtable.row import FamilyNameRegexFilter
-        from gcloud.bigtable.row import RowFilterChain
+        from gcloud.bigtable.row_filters import ColumnQualifierRegexFilter
+        from gcloud.bigtable.row_filters import FamilyNameRegexFilter
+        from gcloud.bigtable.row_filters import RowFilterChain
 
         if col_fam is None:
             col_fam = 'cf1'
@@ -1282,7 +1282,7 @@ class Test__filter_chain_helper(unittest2.TestCase):
                             col_fam=u'cfU', qual=u'qualN')
 
     def test_with_versions(self):
-        from gcloud.bigtable.row import CellsColumnLimitFilter
+        from gcloud.bigtable.row_filters import CellsColumnLimitFilter
 
         versions = 11
         result = self._column_helper(num_filters=3, versions=versions)
@@ -1295,8 +1295,8 @@ class Test__filter_chain_helper(unittest2.TestCase):
 
     def test_with_timestamp(self):
         from gcloud._helpers import _datetime_from_microseconds
-        from gcloud.bigtable.row import TimestampRange
-        from gcloud.bigtable.row import TimestampRangeFilter
+        from gcloud.bigtable.row_filters import TimestampRange
+        from gcloud.bigtable.row_filters import TimestampRangeFilter
 
         timestamp = 1441928298571
         result = self._column_helper(num_filters=3, timestamp=timestamp)
@@ -1330,7 +1330,7 @@ class Test__columns_filter_helper(unittest2.TestCase):
             self._callFUT(columns)
 
     def test_single_column(self):
-        from gcloud.bigtable.row import FamilyNameRegexFilter
+        from gcloud.bigtable.row_filters import FamilyNameRegexFilter
 
         col_fam = 'cf1'
         columns = [col_fam]
@@ -1339,10 +1339,10 @@ class Test__columns_filter_helper(unittest2.TestCase):
         self.assertEqual(result, expected_result)
 
     def test_column_and_column_families(self):
-        from gcloud.bigtable.row import ColumnQualifierRegexFilter
-        from gcloud.bigtable.row import FamilyNameRegexFilter
-        from gcloud.bigtable.row import RowFilterChain
-        from gcloud.bigtable.row import RowFilterUnion
+        from gcloud.bigtable.row_filters import ColumnQualifierRegexFilter
+        from gcloud.bigtable.row_filters import FamilyNameRegexFilter
+        from gcloud.bigtable.row_filters import RowFilterChain
+        from gcloud.bigtable.row_filters import RowFilterUnion
 
         col_fam1 = 'cf1'
         col_fam2 = 'cf2'
@@ -1378,7 +1378,7 @@ class Test__row_keys_filter_helper(unittest2.TestCase):
             self._callFUT(row_keys)
 
     def test_single_row(self):
-        from gcloud.bigtable.row import RowKeyRegexFilter
+        from gcloud.bigtable.row_filters import RowKeyRegexFilter
 
         row_key = b'row-key'
         row_keys = [row_key]
@@ -1387,8 +1387,8 @@ class Test__row_keys_filter_helper(unittest2.TestCase):
         self.assertEqual(result, expected_result)
 
     def test_many_rows(self):
-        from gcloud.bigtable.row import RowFilterUnion
-        from gcloud.bigtable.row import RowKeyRegexFilter
+        from gcloud.bigtable.row_filters import RowFilterUnion
+        from gcloud.bigtable.row_filters import RowKeyRegexFilter
 
         row_key1 = b'row-key1'
         row_key2 = b'row-key2'
