+++ scripts/run_pylint.py
@@ -66,7 +66,7 @@ TEST_RC_ADDITIONS = {
 }
 TEST_RC_REPLACEMENTS = {
     'FORMAT': {
-        'max-module-lines': 1825,
+        'max-module-lines': 1700,
     },
 }
